Split out core framework components
