+++ src/routes/index.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type { JourneyType } from '$lib/types/journey';
 	import { getContext } from 'svelte';
-	import Page from '$lib/components/Page.svelte';
+	import Page from '$lib/components/Framework/Page.svelte';
 
 	const journey: JourneyType = getContext('journey');
 </script>
